requirements/tests.txt
flake8
django-discover-runner
coverage
django-storages
django-celery
boto
django-appconf
